Restructuration.
