+++ livres/templates/auteurs/creer_auteur.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}Créer Auteur{% endblock %}
 {% block content %}
 {% if form.errors %}
