+++ livres/templates/auteurs/detail_auteur.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 
 {% block title %}{{ auteur.nom }}{% endblock %}
