+++ livres/templates/auteurs/modifier_auteur.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}{{ auteur.nom }}{% endblock %}
 {% block content %}
 {% if form.errors %}
