+++ livres/templates/base.html
@@ -287,7 +287,7 @@
         <header>
             <nav class="navbar navbar-expand-lg">
                 <div class="container-fluid px-4">
-                    <a class="navbar-brand" href="{% url 'hub' %}">
+                    <a class="navbar-brand" href="{% url 'livres:rechercher' %}">
                         <img src="{% static 'icons/house.svg' %}" alt="" width="24" height="24">
                         <span>Médiathèque</span>
                     </a>
@@ -302,7 +302,7 @@
                         <ul class="navbar-nav me-auto">
                             {% if perms.auth.view_group %}
                             <li class="nav-item">
-                                <a class="nav-link" href="{% url 'group_list' %}">
+                                <a class="nav-link" href="{% url 'livres:group_list' %}">
                                     <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-people me-1" viewBox="0 0 16 16">
                                         <path d="M15 14s1 0 1-1-1-4-5-4-5 3-5 4 1 1 1 1zm-7.978-1L7 12.996c.001-.264.167-1.03.76-1.72C8.312 10.629 9.282 10 11 10c1.717 0 2.687.63 3.24 1.276.593.69.758 1.457.76 1.72l-.008.002-.014.002zM11 7a2 2 0 1 0 0-4 2 2 0 0 0 0 4m3-2a3 3 0 1 1-6 0 3 3 0 0 1 6 0M6.936 9.28a6 6 0 0 0-1.23-.247A7 7 0 0 0 5 9c-4 0-5 3-5 4q0 1 1 1h4.216A2.24 2.24 0 0 1 5 13c0-1.01.377-2.042 1.09-2.904.243-.294.526-.569.846-.816M4.92 10A5.5 5.5 0 0 0 4 13H1c0-.26.164-1.03.76-1.724.545-.636 1.492-1.256 3.16-1.275ZM1.5 5.5a3 3 0 1 1 6 0 3 3 0 0 1-6 0m3-2a2 2 0 1 0 0 4 2 2 0 0 0 0-4"/>
                                     </svg>
@@ -361,7 +361,7 @@
                         <ul class="navbar-nav ms-auto align-items-center">
                             {% if user.is_authenticated %}
                                 <li class="nav-item me-3">
-                                    <a class="nav-link" href="{% url 'account' %}">
+                                    <a class="nav-link" href="{% url 'livres:account' %}">
                                         <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-person-circle me-1" viewBox="0 0 16 16">
                                             <path d="M11 6a3 3 0 1 1-6 0 3 3 0 0 1 6 0"/>
                                             <path fill-rule="evenodd" d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8m8-7a7 7 0 0 0-5.468 11.37C3.242 11.226 4.805 10 8 10s4.757 1.225 5.468 2.37A7 7 0 0 0 8 1"/>
@@ -370,7 +370,7 @@
                                     </a>
                                 </li>
                                 <li class="nav-item">
-                                    <form method="POST" action="{% url 'logout' %}" class="d-inline">
+                                    <form method="POST" action="{% url 'livres:logout' %}" class="d-inline">
                                         {% csrf_token %}
                                         <button type="submit" class="btn btn-outline-danger btn-sm">
                                             <svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-box-arrow-right me-1" viewBox="0 0 16 16">
@@ -383,12 +383,12 @@
                                 </li>
                             {% else %}
                                 <li class="nav-item me-2">
-                                    <a class="btn btn-outline-primary btn-sm" href="{% url 'login' %}">
+                                    <a class="btn btn-outline-primary btn-sm" href="{% url 'livres:login' %}">
                                         Connexion
                                     </a>
                                 </li>
                                 <li class="nav-item">
-                                    <a class="btn btn-primary btn-sm" href="{% url 'signup' %}">
+                                    <a class="btn btn-primary btn-sm" href="{% url 'livres:signup' %}">
                                         Inscription
                                     </a>
                                 </li>
