+++ livres/templates/lectures/bibliotheque.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}Ma bibliothèque{% endblock %}
