+++ livres/templates/lectures/detail_lecture.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}Détail de la lecture : {{ lecture.livre.nom }}{% endblock %}
