+++ livres/templates/lectures/modifier_lecture.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}Modifier la lecture : {{ lecture.livre.nom }}{% endblock %}
