+++ livres/templates/liste_de_souhaits/liste_de_souhaits.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}Liste de souhaits{% endblock %}
