+++ livres/templates/livres/creer_livre.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}Créer livre{% endblock %}
 {% block content %}
 {% if form.errors %}
