+++ livres/templates/livres/detail_livre.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}{{ livre.nom }}{% endblock %}
