+++ livres/templates/livres/liste_livres.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 
 {% block title %}Rechercher des livres{% endblock %}
