+++ livres/templates/livres/modifier_livre.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}{{ livre.nom }}{% endblock %}
 {% block content %}
 {% if form.errors %}
