+++ livres/templates/tags/creer_tag.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}Créer Tag{% endblock %}
 {% block content %}
 {% if form.errors %}
