+++ livres/templates/tags/liste_tags.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% load static %}
 {% block title %}Rechercher{% endblock %}
 {% block styles %}
