+++ livres/templates/tags/modifier_tag.html
@@ -1,4 +1,4 @@
-{% extends 'base_livres.html' %}
+{% extends 'base.html' %}
 {% block title %}{{ tag.tag }}{% endblock %}
 {% block content %}
 {% if form.errors %}
