+++ livres/templates/user/modifier_account.html
@@ -2,7 +2,7 @@
 {% load static %}
 
 {% block title %}
-Modifier le mot de passe
+Modifier ses informations
 {% endblock %}
 
 {% block content %}
@@ -15,11 +15,11 @@ Modifier le mot de passe
 </div>
 {% endif %}
     <div class="container">
-        <h2>Modifier le mot de passe</h2>
+        <h2>Modifier ses informations</h2>
         <form method="POST">
             {% csrf_token %}
             {{ form.as_p }}
-            <a href="{% url 'account' %}" class="btn btn-outline-secondary">Annuler</a>
+            <a href="{% url 'livres:account' %}" class="btn btn-outline-secondary">Annuler</a>
             <button type="submit" class="btn btn-success">Enregistrer</button>
         </form>
     </div>
