--- templates/hub.html
@@ -1,101 +1,0 @@
-{% extends 'base.html' %}
-{% load static %}
-
-{% block title %}
-Médiathèque
-{% endblock %}
-
-{% block scripts %}
-<script src="{% static 'node_modules/bootstrap/dist/js/bootstrap.bundle.min.js' %}"></script>
-{% endblock %}
-
-{% block styles %}
-    <style>
-        body {
-            background-color: #f8f9fa;
-        }
-        main {
-            padding: 3rem 1rem;
-        }
-        .card:hover {
-            transform: scale(1.05);
-            box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-            transition: transform 0.3s ease, box-shadow 0.3s ease;
-        }
-        .card img {
-            height: 150px;
-            object-fit: cover;
-        }
-        h1 {
-            color: #343a40;
-            font-size: 2.5rem;
-            margin-bottom: 2rem;
-            text-align: center;
-        }
-        p {
-            text-align: center;
-            font-size: 1.2rem;
-            color: #6c757d;
-        }
-        .card-group {
-            display: flex;
-            flex-wrap: wrap;
-            justify-content: center;
-            gap: 1.5rem;
-        }
-        .list-group a {
-            text-decoration: none;
-            color: #007bff;
-        }
-        .list-group a:hover {
-            text-decoration: underline;
-        }
-    </style>
-{% endblock %}
-
-{% block content %}
-    <div class="container">
-        <h1>Bienvenue sur le site de gestion de votre médiathèque !</h1>
-        <p>Pour continuer, choisissez une catégorie que vous souhaitez consulter.</p>
-        <div class="card-group">
-            <a href="{% url 'livres:rechercher' %}" class="text-decoration-none">
-                <div class="card text-center">
-                    {% if livre %}
-                    <img src="{{ livre.image.url }}" alt="Livres" class="card-img-top">
-                    {% else %}
-                    <img src="{% static 'default.png' %}" alt="Livres" class="card-img-top">
-                    {% endif %}
-                    <div class="card-body">
-                        <h5 class="card-title">Livres</h5>
-                        <p class="card-text">Romans, BD, Comics, Mangas, ...</p>
-                    </div>
-                </div>
-            </a>
-            <a href="{% url 'schema-swagger-ui' %}" class="text-decoration-none">
-                <div class="card text-center">
-                    <img src="{% static 'documentation_api.png' %}" alt="Livres" class="card-img-top">
-                    <div class="card-body">
-                        <h5 class="card-title">Documentation API</h5>
-                        <p class="card-text">Consultez les spécifications de l'API.</p>
-                    </div>
-                </div>
-            </a>
-            <a href="#" class="text-decoration-none">
-                <div class="card text-center">
-                    <div class="card-body">
-                        <h5 class="card-title">Animés/Séries/Films</h5>
-                        <p class="card-text">Découvrez vos médias préférés.</p>
-                    </div>
-                </div>
-            </a>
-            <a href="#" class="text-decoration-none">
-                <div class="card text-center">
-                    <div class="card-body">
-                        <h5 class="card-title">DVD/Bluray</h5>
-                        <p class="card-text">Collectionnez vos classiques.</p>
-                    </div>
-                </div>
-            </a>
-        </div>
-    </div>
-{% endblock %}
